+++ lib/src/platforms/android/android.dart
@@ -2,3 +2,4 @@ export 'nearby_android_service.dart';
 export 'nearby_service_android_interface.dart';
 export 'model/nearby_connection_info.dart';
 export 'model/nearby_device.dart';
+export 'utils/exception.dart';