+++ lib/src/utils/exception.dart
@@ -14,33 +14,87 @@ class NearbyServiceException implements Exception {
   }
 
   ///
-  /// A call from an unsupported platform.
+  /// Usage of the plugin on an unsupported platform
   ///
-  factory NearbyServiceException.unsupportedPlatform({required String caller}) {
-    return NearbyServiceException(
-      '$caller is not supported for platform ${Platform.operatingSystem}',
-    );
-  }
+  factory NearbyServiceException.unsupportedPlatform({
+    required String caller,
+  }) =>
+      NearbyServiceUnsupportedPlatformException(caller: caller);
 
   ///
-  /// A decoding error.
+  /// Error decoding messages from native platform to Dart (open an issue if
+  /// this happens!)
   ///
-  factory NearbyServiceException.unsupportedDecoding(dynamic value) {
-    return NearbyServiceException(
-      'Got unknown value=$value with runtimeType=${value.runtimeType}',
-    );
-  }
+  factory NearbyServiceException.unsupportedDecoding(dynamic value) =>
+      NearbyServiceUnsupportedDecodingException(value);
 
-  factory NearbyServiceException.invalidMessage(NearbyMessageContent content) {
-    return NearbyServiceException(
-      'The message="$content" is not valid',
-    );
-  }
+  ///
+  /// An attempt to send an invalid message on the sender's side. Add content
+  /// validation to your messages
+  ///
+  factory NearbyServiceException.invalidMessage(NearbyMessageContent content) =>
+      NearbyServiceInvalidMessageException(content);
 
   final Object? error;
 
   @override
   String toString() {
     return 'NearbyServiceException{error: $error}';
+  }
+}
+
+///
+/// Usage of the plugin on an unsupported platform
+///
+class NearbyServiceUnsupportedPlatformException extends NearbyServiceException {
+  ///
+  /// Usage of the plugin on an unsupported platform - default constructor
+  ///
+  NearbyServiceUnsupportedPlatformException({required String caller})
+      : super(
+          '$caller is not supported for platform ${Platform.operatingSystem}',
+        );
+
+  @override
+  String toString() {
+    return 'NearbyServiceUnsupportedPlatformException{error: $error}';
+  }
+}
+
+///
+/// Error decoding messages from native platform to Dart (open an issue if
+/// this happens!)
+///
+class NearbyServiceUnsupportedDecodingException extends NearbyServiceException {
+  ///
+  /// A decoding error - default constructor
+  ///
+  NearbyServiceUnsupportedDecodingException(dynamic value)
+      : super(
+          'Got unknown value=$value with runtimeType=${value.runtimeType}',
+        );
+
+  @override
+  String toString() {
+    return 'NearbyServiceUnsupportedDecodingException{error: $error}';
+  }
+}
+
+///
+/// An attempt to send an invalid message on the sender's side. Add content
+/// validation to your messages
+///
+class NearbyServiceInvalidMessageException extends NearbyServiceException {
+  ///
+  /// Invalid message error - default constructor
+  ///
+  NearbyServiceInvalidMessageException(NearbyMessageContent content)
+      : super(
+          'The message="$content" is not valid',
+        );
+
+  @override
+  String toString() {
+    return 'NearbyServiceInvalidMessageException{error: $error}';
   }
 }