+++ lib/src/utils/file_socket.dart
@@ -62,10 +62,10 @@ class FilesSocket {
       addChunk(event);
     } else if (event == separateCommandOf(_currentFileIndex)) {
       _futures.add(_createFile(_currentFileIndex));
+      Logger.info('Completed receiving file №${_currentFileIndex + 1}');
       _currentFileIndex = _currentFileIndex + 1;
       _chunksCount = 0;
       _bytesTable['$_currentFileIndex'] = [];
-      Logger.info('Completed receiving file №${_currentFileIndex - 1}');
     } else if (event == finishCommand) {
       await Future.wait(_futures);
       Logger.info('Files pack ${filesRequest.id} was created');